Use interpolated string so the windows_path resource doesn't modify the home attribute value.
For some unknown reason the Chef env resource temporarily modifies the node['seven_zip']['home'] attribute during the initial Chef run that installs and configures 7-zip. Instead of using the default value it overrides it with the new PATH value. On the next run the value is fine. Very odd.